refactor: make domain a leaf layer by relocating shared value types (#495)
The intended dependency arrow is app -> domain, but three domain modules pointed
back at app:: (and one at ui::). Move the pure value types they needed into
domain so the layer stops depending on its dependents:

- SendRequest -> new domain/send.rs
- VisibleImage, ImageRenderResult, LinkRegion -> domain/image.rs
- GroupMenuState, PinPending, PollVotePending -> domain/overlays.rs

domain/image.rs, domain/pending.rs, and domain/overlays.rs now import these from
within domain instead of from app::/ui::. app.rs re-exports the relocated types
(pub use crate::domain::...) and ui re-exports LinkRegion, so every existing
crate::app::* / crate::ui::LinkRegion reference across the handlers, UI, and main
loop keeps resolving unchanged. No behavior change.

The only remaining domain -> app reference is a doc-comment link. This unblocks
the rest of the #352 extraction work.
