+++ src/domain/mod.rs
@@ -18,11 +18,12 @@ mod pending;
 mod reaction;
 mod scroll;
 mod search;
+mod send;
 mod typing;
 
 pub use emoji_picker::{CATEGORIES, EmojiPickerAction, EmojiPickerSource, EmojiPickerState};
 pub use file_picker::{FilePickerOutcome, FilePickerState};
-pub use image::{ImageMode, ImageState};
+pub use image::{ImageMode, ImageRenderResult, ImageState, LinkRegion, VisibleImage};
 pub use input::InputState;
 pub use lock::{LockPhase, LockState};
 pub use lock::{hash_passphrase, load_hash, lock_hash_path, save_hash, verify_passphrase};
@@ -30,11 +31,13 @@ pub use mouse::MouseState;
 pub use notification::{NotificationPreview, NotificationState};
 pub use overlays::{
     ActionMenuState, ContactsOverlayState, ForwardOverlayState, GroupMenuOverlayState,
-    KeybindingsOverlayState, PinDurationOverlayState, PollVoteOverlayState, ProfileOverlayState,
-    SettingsOverlayState, SettingsProfileOverlayState, ThemePickerState, VerifyOverlayState,
+    GroupMenuState, KeybindingsOverlayState, PinDurationOverlayState, PinPending,
+    PollVoteOverlayState, PollVotePending, ProfileOverlayState, SettingsOverlayState,
+    SettingsProfileOverlayState, ThemePickerState, VerifyOverlayState,
 };
 pub use pending::{BufferedReceipt, PendingState};
 pub use reaction::ReactionState;
 pub use scroll::ScrollState;
 pub use search::{SearchAction, SearchState};
+pub use send::SendRequest;
 pub use typing::TypingState;
